remove all helpers on src/helpers
